requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pymupdf==1.23.8
numpy==1.24.3
python-dotenv==1.0.0
requests==2.31.0
